ENORME PUSH

+++ maths/ft_maths.h
@@ -27,7 +27,7 @@ float	ft_sqrt(float nb);
 float	ft_div(float a, float b);
 float	ft_root(float nb, float root);
 float	ft_pow2(float nb, float pow);
-double	ft_pow(double base, double power);
+int	ft_pow(int base, int power);
 double	ft_pow3(double base, double power);
 int		is_prime(int nb);
 void	prime_factors(int nb);
